Create 2022-03-12-title.md
Erelabs is a Blockchain technology start-up company that employs the dynamism of metaverse to advance technology advancement for the human race. We are a group of creative technologists that are passionate, skilled, and curious about the technology in focus. We believe that simplicity is genius. That is why we are developing a metaverse that is easy to use for all and sundry. We call it Plug & Play. Despite our fundamental belief in making this technology simple, we will make it a holistic one. We are going to embed and use all the pillars necessary for running an efficient metaverse on the blockchain. Pillars such as; Augmented Reality, Virtual Reality, 3D reconstruction, Artificial intelligence, the Internet of Things, Blockchain, and Cryptocurrency will be used to develop a one-of-a-kind metaverse that is easy to use. Motto: Plug X Play